chore: restructure the app
