+++ module.d.ts
@@ -1,7 +1,7 @@
 declare namespace NodeJS {
     export interface ProcessEnv {
-        PORT: string;
+        PORT: string | undefined;
         MONGODB_API_KEY: string;
-        RANDOM_TOKEN_SECRET: string;
+        RANDOM_TOKEN_SECRET: 'Secret | GetPublicKeyOrSecret';
     }
 }
