+++ src/echoStore/v1/models/user.model.ts
